etl/requirements.txt
pandas>=1.5.0
numpy>=1.21.0
python-dateutil>=2.8.2
pytz>=2021.3
psycopg2-binary>=2.9.5
pymongo>=4.3.3
sqlalchemy>=1.4.41
python-dotenv>=0.21.0 